tests: refactor droplet test into own program
Previously the droplet tests were integrated into the sd_backend tests.
With the new backend loading, we cannot configure a droplet backend if
we did not build it, thus we need to extract the droplet tests so we can
disable them completely.
