+++ core/src/tests/configs/droplet_backend/bareos-sd.d/device/droplet.conf.in
@@ -2,7 +2,7 @@ Device {
   Name = droplet
   Media Type = S3
   Device Type = droplet
-  Device Options = "profile=@CMAKE_CURRENT_SOURCE_DIR@/src/tests/configs/sd_backend/droplet.profile,bucket=bareos-test,chunksize=10M"
+  Device Options = "profile=@CMAKE_CURRENT_SOURCE_DIR@/src/tests/configs/droplet_backend/droplet.profile,bucket=bareos-test,chunksize=10M"
   Maximum Concurrent Jobs = 1
   Archive Device = "Cloud Storage"
   LabelMedia = yes
