+++ core/src/tests/configs/droplet_backend/bareos-sd.d/storage/myself.conf.in
@@ -1,0 +1,5 @@
+Storage {
+  Name = test-sd
+  @UNCOMMENT_SD_BACKEND_DIRECTORY@Backend Directory = @PROJECT_BINARY_DIR@/src/stored/backends
+  Working Directory = @PROJECT_BINARY_DIR@/
+}
